remove RepositorySettingsStore in favour of plain functions
