+++ app/src/lib/git/index.ts
@@ -31,3 +31,4 @@ export * from './rm'
 export * from './mergetool'
 export * from './submodule'
 export * from './interpret-trailers'
+export * from './gitignore'
