+++ app/src/lib/stores/app-store.ts
@@ -11,7 +11,6 @@ import {
   IssuesStore,
   PullRequestStore,
   RepositoriesStore,
-  RepositorySettingsStore,
   SignInStore,
 } from '.'
 import { Account } from '../../models/account'
@@ -120,6 +119,9 @@ import {
   push as pushRepo,
   renameBranch,
   updateRef,
+  saveGitIgnore,
+  readGitIgnoreAtRoot,
+  appendIgnoreRule,
 } from '../git'
 import { IGitAccount } from '../git/authentication'
 import {
@@ -215,10 +217,6 @@ export class AppStore extends TypedBaseStore<IAppState> {
 
   /** GitStores keyed by their hash. */
   private readonly gitStores = new Map<string, GitStore>()
-  private readonly repositorySettingsStores = new Map<
-    string,
-    RepositorySettingsStore
-  >()
 
   private readonly localRepositoryStateLookup = new Map<
     number,
@@ -524,30 +522,6 @@ export class AppStore extends TypedBaseStore<IAppState> {
     }
 
     return gitStore
-  }
-
-  private removeRepositorySettingsStore(repository: Repository) {
-    const key = repository.hash
-
-    if (this.repositorySettingsStores.has(key)) {
-      this.repositorySettingsStores.delete(key)
-    }
-  }
-
-  private getRepositorySettingsStore(
-    repository: Repository
-  ): RepositorySettingsStore {
-    let store = this.repositorySettingsStores.get(repository.hash)
-
-    if (store == null) {
-      store = new RepositorySettingsStore(repository)
-
-      store.onDidError(error => this.emitError(error))
-
-      this.repositorySettingsStores.set(repository.hash, store)
-    }
-
-    return store
   }
 
   private clearSelectedCommit(repository: Repository) {
@@ -1090,7 +1064,6 @@ export class AppStore extends TypedBaseStore<IAppState> {
       // ensures we don't accidentally run any Git operations against the
       // wrong location if the user then relocates the `.git` folder elsewhere
       this.removeGitStore(repository)
-      this.removeRepositorySettingsStore(repository)
       return Promise.resolve(null)
     }
 
@@ -3056,14 +3029,12 @@ export class AppStore extends TypedBaseStore<IAppState> {
     repository: Repository,
     text: string
   ): Promise<void> {
-    const repositorySettingsStore = this.getRepositorySettingsStore(repository)
-    return repositorySettingsStore.saveGitIgnore(text)
+    return saveGitIgnore(repository, text)
   }
 
   /** This shouldn't be called directly. See `Dispatcher`. */
   public async _readGitIgnore(repository: Repository): Promise<string | null> {
-    const repositorySettingsStore = this.getRepositorySettingsStore(repository)
-    return repositorySettingsStore.readGitIgnore()
+    return readGitIgnoreAtRoot(repository)
   }
 
   /** Has the user opted out of stats reporting? */
@@ -3155,9 +3126,7 @@ export class AppStore extends TypedBaseStore<IAppState> {
     repository: Repository,
     pattern: string | string[]
   ): Promise<void> {
-    const repoSettingsStore = this.getRepositorySettingsStore(repository)
-
-    await repoSettingsStore.ignore(pattern)
+    await appendIgnoreRule(repository, pattern)
 
     return this._refreshRepository(repository)
   }
