+++ app/src/lib/stores/index.ts
@@ -9,5 +9,4 @@ export * from './repositories-store'
 export * from './sign-in-store'
 export * from './token-store'
 export * from './pull-request-store'
-export * from './repository-settings-store'
 export { UpstreamRemoteName } from './helpers/find-upstream-remote'
